+++ src/util/CrashHandler.h
@@ -9,7 +9,7 @@ class CrashHandler : public urchin::SignalReceptor {
     public:
         void onLogContainFailure();
         void onException(const std::exception& e);
-        void onSignalReceived(int) override;
+        void onSignalReceived(unsigned long) override;
 };
 
 #endif
